+++ lib/workers.rb
@@ -1,8 +1,8 @@
 require 'thread'
 require 'monitor'
 require 'set'
-require 'sorted_set'
 
+require 'workers/sorted_set'
 require 'workers/version'
 require 'workers/exceptions'
 require 'workers/helpers'
